refactor:更新PublishFeedbackRequest中的字段注解，重命名SecurityHeaderConfig包，优化UserServiceImpl中的异常处理
