+++ bigproject/src/main/java/org/bigseven/security/SecurityHeaderConfig.java
@@ -1,4 +1,4 @@
-package org.bigseven.config;
+package org.bigseven.security;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
